+++ admin/sidebar.php
@@ -442,6 +442,18 @@
               </a>
             </li>
             <li class="sidebar-item">
+              <a class="sidebar-link" href="billing" aria-expanded="false">
+                <i class="ti ti-receipt"></i>
+                <span class="hide-menu">Billing</span>
+              </a>
+            </li>
+            <li class="sidebar-item">
+              <a class="sidebar-link" href="workflow" aria-expanded="false">
+                <i class="ti ti-timeline"></i>
+                <span class="hide-menu">Workflow Timeline</span>
+              </a>
+            </li>
+            <li class="sidebar-item">
               <a class="sidebar-link" href="segment" aria-expanded="false">
                 <i class="ti ti-list-details"></i>
                 <span class="hide-menu">Segment/location</span>